Serialization.
