+++ gui/executionGraphGui/messages/schemas/ts/GraphSerializationMessages_generated.ts
@@ -1,0 +1,383 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+
+import * as NS12623504695714931604 from "@eg/serialization/ExecutionGraph_generated";
+/**
+ * @constructor
+ */
+export namespace executionGraphGui.serialization{
+export class LoadGraphRequest {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns LoadGraphRequest
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):LoadGraphRequest {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param LoadGraphRequest= obj
+ * @returns LoadGraphRequest
+ */
+static getRoot(bb:flatbuffers.ByteBuffer, obj?:LoadGraphRequest):LoadGraphRequest {
+  return (obj || new LoadGraphRequest).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+filePath():string|null
+filePath(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+filePath(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static start(builder:flatbuffers.Builder) {
+  builder.startObject(1);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset filePathOffset
+ */
+static addFilePath(builder:flatbuffers.Builder, filePathOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, filePathOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static end(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  builder.requiredField(offset, 4); // filePath
+  return offset;
+};
+
+static create(builder:flatbuffers.Builder, filePathOffset:flatbuffers.Offset):flatbuffers.Offset {
+  LoadGraphRequest.start(builder);
+  LoadGraphRequest.addFilePath(builder, filePathOffset);
+  return LoadGraphRequest.end(builder);
+}
+}
+}
+/**
+ * @constructor
+ */
+export namespace executionGraphGui.serialization{
+export class LoadGraphResponse {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns LoadGraphResponse
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):LoadGraphResponse {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param LoadGraphResponse= obj
+ * @returns LoadGraphResponse
+ */
+static getRoot(bb:flatbuffers.ByteBuffer, obj?:LoadGraphResponse):LoadGraphResponse {
+  return (obj || new LoadGraphResponse).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+graphId():string|null
+graphId(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+graphId(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param executionGraph.serialization.ExecutionGraph= obj
+ * @returns executionGraph.serialization.ExecutionGraph|null
+ */
+graph(obj?:NS12623504695714931604.executionGraph.serialization.ExecutionGraph):NS12623504695714931604.executionGraph.serialization.ExecutionGraph|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? (obj || new NS12623504695714931604.executionGraph.serialization.ExecutionGraph).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param number index
+ * @returns number
+ */
+visualization(index: number):number|null {
+  var offset = this.bb!.__offset(this.bb_pos, 8);
+  return offset ? this.bb!.readUint8(this.bb!.__vector(this.bb_pos + offset) + index) : 0;
+};
+
+/**
+ * @returns number
+ */
+visualizationLength():number {
+  var offset = this.bb!.__offset(this.bb_pos, 8);
+  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
+};
+
+/**
+ * @returns Uint8Array
+ */
+visualizationArray():Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 8);
+  return offset ? new Uint8Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static start(builder:flatbuffers.Builder) {
+  builder.startObject(3);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset graphIdOffset
+ */
+static addGraphId(builder:flatbuffers.Builder, graphIdOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, graphIdOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset graphOffset
+ */
+static addGraph(builder:flatbuffers.Builder, graphOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, graphOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset visualizationOffset
+ */
+static addVisualization(builder:flatbuffers.Builder, visualizationOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(2, visualizationOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param Array.<number> data
+ * @returns flatbuffers.Offset
+ */
+static createVisualizationVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
+  builder.startVector(1, data.length, 1);
+  for (var i = data.length - 1; i >= 0; i--) {
+    builder.addInt8(data[i]);
+  }
+  return builder.endVector();
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number numElems
+ */
+static startVisualizationVector(builder:flatbuffers.Builder, numElems:number) {
+  builder.startVector(1, numElems, 1);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static end(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  builder.requiredField(offset, 4); // graphId
+  builder.requiredField(offset, 6); // graph
+  return offset;
+};
+
+static create(builder:flatbuffers.Builder, graphIdOffset:flatbuffers.Offset, graphOffset:flatbuffers.Offset, visualizationOffset:flatbuffers.Offset):flatbuffers.Offset {
+  LoadGraphResponse.start(builder);
+  LoadGraphResponse.addGraphId(builder, graphIdOffset);
+  LoadGraphResponse.addGraph(builder, graphOffset);
+  LoadGraphResponse.addVisualization(builder, visualizationOffset);
+  return LoadGraphResponse.end(builder);
+}
+}
+}
+/**
+ * @constructor
+ */
+export namespace executionGraphGui.serialization{
+export class SaveGraphRequest {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns SaveGraphRequest
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):SaveGraphRequest {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param SaveGraphRequest= obj
+ * @returns SaveGraphRequest
+ */
+static getRoot(bb:flatbuffers.ByteBuffer, obj?:SaveGraphRequest):SaveGraphRequest {
+  return (obj || new SaveGraphRequest).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+graphId():string|null
+graphId(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+graphId(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+filePath():string|null
+filePath(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+filePath(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @returns boolean
+ */
+overwrite():boolean {
+  var offset = this.bb!.__offset(this.bb_pos, 8);
+  return offset ? !!this.bb!.readInt8(this.bb_pos + offset) : false;
+};
+
+/**
+ * @param number index
+ * @returns number
+ */
+visualization(index: number):number|null {
+  var offset = this.bb!.__offset(this.bb_pos, 10);
+  return offset ? this.bb!.readUint8(this.bb!.__vector(this.bb_pos + offset) + index) : 0;
+};
+
+/**
+ * @returns number
+ */
+visualizationLength():number {
+  var offset = this.bb!.__offset(this.bb_pos, 10);
+  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
+};
+
+/**
+ * @returns Uint8Array
+ */
+visualizationArray():Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 10);
+  return offset ? new Uint8Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static start(builder:flatbuffers.Builder) {
+  builder.startObject(4);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset graphIdOffset
+ */
+static addGraphId(builder:flatbuffers.Builder, graphIdOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, graphIdOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset filePathOffset
+ */
+static addFilePath(builder:flatbuffers.Builder, filePathOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, filePathOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param boolean overwrite
+ */
+static addOverwrite(builder:flatbuffers.Builder, overwrite:boolean) {
+  builder.addFieldInt8(2, +overwrite, +false);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset visualizationOffset
+ */
+static addVisualization(builder:flatbuffers.Builder, visualizationOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(3, visualizationOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param Array.<number> data
+ * @returns flatbuffers.Offset
+ */
+static createVisualizationVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
+  builder.startVector(1, data.length, 1);
+  for (var i = data.length - 1; i >= 0; i--) {
+    builder.addInt8(data[i]);
+  }
+  return builder.endVector();
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number numElems
+ */
+static startVisualizationVector(builder:flatbuffers.Builder, numElems:number) {
+  builder.startVector(1, numElems, 1);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static end(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  builder.requiredField(offset, 4); // graphId
+  builder.requiredField(offset, 6); // filePath
+  return offset;
+};
+
+static create(builder:flatbuffers.Builder, graphIdOffset:flatbuffers.Offset, filePathOffset:flatbuffers.Offset, overwrite:boolean, visualizationOffset:flatbuffers.Offset):flatbuffers.Offset {
+  SaveGraphRequest.start(builder);
+  SaveGraphRequest.addGraphId(builder, graphIdOffset);
+  SaveGraphRequest.addFilePath(builder, filePathOffset);
+  SaveGraphRequest.addOverwrite(builder, overwrite);
+  SaveGraphRequest.addVisualization(builder, visualizationOffset);
+  return SaveGraphRequest.end(builder);
+}
+}
+}
